requirements.txt
altgraph==0.17.4
packaging==23.2
pygame-ce==2.4.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
setuptools==69.1.1
